캐릭터 선택 및 덱 설정 UI, 인게임에서 캐릭터 선택 및 업그래이드 등 인터페이스 화

+++ CharacterController/EnemyController.cs
@@ -153,7 +153,7 @@ public class EnemyController : MonoBehaviour, IGamePlayCharacter
     /// 외부(공격 유닛 등)에서 호출하는 피격 메서드입니다.
     /// </summary>
     /// <param name="atk">공격력</param>
-    public virtual void Hit(int atk)
+    public virtual void Hit(float atk)
     {
         if (isDie) return;
